+++ src/commands/hardis/work/refresh.ts
@@ -53,23 +53,30 @@ export default class RefreshTask extends SfdxCommand {
         await execCommand(pullCommand, this, { output: true, fail: true });
 
         // Stash
-        uxLog(this, c.cyan(`Stashing your local updates before merging ${c.green(config.developmentBranch)} into your local branch ${c.green(localBranch)}...`));
-        await git({output:true}).stash();
-
+        uxLog(this, c.cyan(`Stashing your uncommited updates in ${c.green(localBranch)} before merging ${c.green(config.developmentBranch)} into your local branch ${c.green(localBranch)}...`));
+        const stashResult = await git({output:true}).stash();
+        const stashed = stashResult.includes('Saved working directory');
         // Pull most recent version of development branch
-        uxLog(this, c.cyan(`Retrieving most recent version of remote branch ${config.developmentBranch}...`));
+        uxLog(this, c.cyan(`Pulling most recent version of remote branch ${c.green(config.developmentBranch)}...`));
         await git({output:true}).fetch();
         await git({output:true}).checkout(config.developmentBranch);
-        await git({output:true}).pull();
-        // Merge into current branch
-        // Create new commit from merge
-        uxLog(this, c.cyan(`Creating a merge commit within ${localBranch}...`));
-        await git({output:true}).mergeFromTo(config.developmentBranch, localBranch)
-            .add('./*')
-            .commit(`[sfdx-hardis] Merge from ${config.developmentBranch}`);
-        uxLog(this, c.cyan(`Restoring stash into your local branch ${c.green(localBranch)}...`));
+        const pullRes = await git({output:true}).pull();
+        // Merge into current branch if necessary
+        if (pullRes.summary.changes > 0) {
+            // Create new commit from merge
+            uxLog(this, c.cyan(`Creating a merge commit of ${c.green(config.developmentBranch)} within ${c.green(localBranch)}...`));
+            await git({output:true}).mergeFromTo(config.developmentBranch, localBranch)
+                .add('./*')
+                .commit(`[sfdx-hardis] Merge updates from ${config.developmentBranch}`);
+        }
+        else {
+            uxLog(this, c.cyan(`Local branch ${c.green(localBranch)} is already up to date with ${c.green(config.developmentBranch)}`));
+        }
         await git({output:true}).checkout(localBranch);
-        await git({output:true}).stash(['pop']);
+        if (stashed) {
+            uxLog(this, c.cyan(`Restoring stash into your local branch ${c.green(localBranch)}...`));
+            await git({output:true}).stash(['pop']);
+        }
 
         // Return an object to be displayed with --json
         return { outputString: 'Refreshed the task' };