+++ src/commands/hardis/work/resetselection.ts
@@ -1,9 +1,9 @@
 /* jscpd:ignore-start */
 import { flags, SfdxCommand } from '@salesforce/command';
-import { Messages } from '@salesforce/core';
+import { Messages, SfdxError } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import * as c from 'chalk';
-import {  uxLog } from '../../../common/utils';
+import {  getCurrentGitBranch, git, uxLog } from '../../../common/utils';
 import { getConfig } from '../../../config';
 
 // Initialize Messages with the current plugin directory
@@ -20,7 +20,7 @@ export default class RebuildSelection extends SfdxCommand {
     public static description = messages.getMessage('rebuildSelection');
 
     public static examples = [ 
-        '$ sfdx hardis:work:rebuildselection'
+        '$ sfdx hardis:work:resetsave'
     ];
 
     // public static args = [{name: 'file'}];
@@ -45,10 +45,24 @@ export default class RebuildSelection extends SfdxCommand {
     public async run(): Promise<AnyJson> {
         this.debugMode = this.flags.debug || false;
         const config = await getConfig('project');
-        //const localBranch = await getCurrentGitBranch();
-        uxLog(this, c.cyan(`This script will rebuild selection that you will want to publish in ${c.green(config.developmentBranch)}`));
+        uxLog(this, c.cyan(`This script will rebuild selection that you will want to publish to ${c.green(config.developmentBranch)}`));
 
+        const currentGitBranch = await getCurrentGitBranch();
+        if (currentGitBranch === config.developmentBranch) {
+            throw new SfdxError(c.red('[sfdx-hardis] You can not revert commits of a protected branch !'))
+        }
+        // List all commits since the branch creation
+        const logResult = await git().log([`${config.developmentBranch}..${currentGitBranch}`]);
+        const commitstoRevert = logResult.all;
+        // Revert commits
+        for (const commit of commitstoRevert) {
+            await git({output:true}).revert(commit.hash,['--no-commit']);
+        }
+
+        uxLog(this,c.cyan(`The following items are not available for selection`));
+        await git({output:true}).status();
+        uxLog(this,c.cyan(`Selection has been reset`));
         // Return an object to be displayed with --json
-        return { outputString: 'Rebuild selection' };
+        return { outputString: 'Reset selection pocessed' };
     }
 }