+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - New command **hardis:org:select**
+- New command **hardis:work:resetselection**
 - **hardis:work:save**: Upgrade package.xml and destructiveChanges.xml from git diff
+- Improve console logging of git operations
 
 ## [1.4.1] 2021-03-03